requirements.txt
numpy==1.23.4
numpy==1.20.1
pandas==1.2.4
Pillow==9.2.0
Pillow==8.2.0
Pillow==10.0.0
scikit_learn==0.24.1
torch==2.0.0+cu117
torchvision==0.15.1+cu117
transformers==4.30.2
